+++ src/elisa/analytics/binary_fit/lc_fit.py
@@ -21,11 +21,12 @@ class LCFit(FitResultHandler):
     """
     Class with common methods used during an LC fit.
     """
-    def __init__(self, morphology, atmosphere_model):
+    def __init__(self, morphology, atmosphere_model, limb_darkening_coefficients):
         super().__init__()
         self.morphology = morphology
         self.fit_method_instance: Union[LCFitLeastSquares, LCFitMCMC, None] = None
         self.atmosphere_model: Union[dict, None] = None
+        self.limb_darkening_coefficients: Union[dict, None] = None
 
     def coefficient_of_determination(self, model_parameters, data, discretization, interp_treshold):
         """
@@ -53,10 +54,11 @@ class LCFitMCMC(LCFit):
     """
     Class for LC fitting using the MCMC method.
     """
-    def __init__(self, morphology, atmosphere_model):
-        super().__init__(morphology, atmosphere_model)
+    def __init__(self, morphology, atmosphere_model, limb_darkening_coefficients):
+        super().__init__(morphology, atmosphere_model, limb_darkening_coefficients)
         self.fit_method_instance = self.resolve_fit_cls(morphology)()
         self.fit_method_instance.atmosphere_model = atmosphere_model
+        self.fit_method_instance.limb_darkening_coefficients = limb_darkening_coefficients
 
         self.flat_chain = None
         self.flat_chain_path = None
@@ -157,10 +159,11 @@ class LCFitLeastSquares(LCFit):
     """
     Class for LC fitting using the Least-Squares method.
     """
-    def __init__(self, morphology, atmosphere_model):
-        super().__init__(morphology, atmosphere_model)
+    def __init__(self, morphology, atmosphere_model, limb_darkening_coefficients):
+        super().__init__(morphology, atmosphere_model, limb_darkening_coefficients)
         self.fit_method_instance = self.resolve_fit_cls(morphology)()
         self.fit_method_instance.atmosphere_model = atmosphere_model
+        self.fit_method_instance.limb_darkening_coefficients = limb_darkening_coefficients
 
     def fit(self, x0: BinaryInitialParameters, data, **kwargs):
         """